Move rules.py -> check_rules.py
